Keep original values for `gtfs_arrival_dt` and `gtfs_departure_dt`
