+++ tests/bus_performance_manager/test_events_metrics.py
@@ -10,6 +10,60 @@ from dataframely.random import Generator
 
 from lamp_py.bus_performance_manager.events_metrics import BusPerformanceMetrics
 
+@pytest.mark.parametrize(
+    ["point_type", "gtfs_last_in_transit_dt", "gtfs_arrival_dt", "num_rows"],
+    [
+        (
+            "end",
+            None,
+            None,
+            nullcontext(3),
+        ),
+        (
+            "end",
+            datetime(2000, 1, 1, 2),
+            None,
+            pytest.raises(ValidationError, match="has_arrival_dt"),
+        ),
+        (
+            "end",
+            datetime(2000, 1, 1, 2),
+            datetime(2000, 1, 1, 2),
+            nullcontext(3),
+        ),
+    ],
+    ids=[
+        "missing-GTFS-stop-data",  # pass
+        "final-stop-but-no-gtfs-arrival",  # fail
+        "all-data-present",  # pass
+    ],
+)
+def test_dy_final_stop_has_arrival_dt(
+    dy_gen: Generator,
+    point_type: str,
+    gtfs_last_in_transit_dt: datetime,
+    gtfs_arrival_dt: list[datetime],
+    num_rows: pytest.RaisesExc,
+) -> None:
+    "It returns false if the last TM stop or, if not available, GTFS stop, has in-transit data but not a gtfs_arrival_dt."
+    df = BusPerformanceMetrics.sample(num_rows=3, generator=dy_gen).with_columns(
+        trip_id=pl.lit("1"),
+        tm_pullout_id=pl.lit("0"),
+        route_id=pl.lit("a"),
+        vehicle_label=pl.lit("x"),
+        service_date=pl.lit(date(2000, 1, 1)),
+        tm_stop_sequence=pl.Series(values=[1, 2, 3]),
+        stop_sequence=pl.Series(values=[1, 2, 3]),
+        point_type=pl.Series(values=["start", "mid", point_type]),
+        stop_departure_dt=pl.lit(None),
+        gtfs_last_in_transit_dt=pl.Series(
+            values=[datetime(2000, 1, 1), datetime(2000, 1, 1, 1), gtfs_last_in_transit_dt]
+        ),
+        gtfs_arrival_dt=pl.Series(values=[datetime(2000, 1, 1), datetime(2000, 1, 1, 1), gtfs_arrival_dt]),
+    )
+
+    with num_rows:
+        assert BusPerformanceMetrics.validate(df, cast=True).height == num_rows.enter_result  # type: ignore[attr-defined]
 
 @pytest.mark.parametrize(
     ["stop_arrival_dt", "stop_departure_dt", "travel_time_seconds", "stopped_duration_seconds", "num_rows"],
